factored out album from app.py
