+++ admin.py
@@ -7,6 +7,7 @@ from app import db
 
 blueprint = Blueprint('admin', __name__, template_folder='templates/admin')
 
+
 @blueprint.route("/")
 @is_admin
 def admin():
@@ -15,18 +16,19 @@ def admin():
         return "admin"
     return "not admin"
 
+
 @blueprint.route("/roles", methods=["GET", "POST"])
 @is_admin
 def roles():
-        if request.method == "POST":
-            user_id = request.form['user_id']
-            user = User.query.filter_by(id=user_id).first()
-            user.is_admin = is_checked(request, "is_admin")
-            user.can_upload = is_checked(request, "can_upload")
-            user.can_view = is_checked(request, "can_view")
-            db.session.add(user)
-            db.session.commit()
-            return render_template("roles_partial.html", user=user)
-        else:
-            all_users = User.query.all()
-            return render_template("roles.html", all_users=all_users)
+    if request.method == "POST":
+        user_id = request.form['user_id']
+        user = User.query.filter_by(id=user_id).first()
+        user.is_admin = is_checked(request, "is_admin")
+        user.can_upload = is_checked(request, "can_upload")
+        user.can_view = is_checked(request, "can_view")
+        db.session.add(user)
+        db.session.commit()
+        return render_template("roles_partial.html", user=user)
+    else:
+        all_users = User.query.all()
+        return render_template("roles.html", all_users=all_users)
