+++ templates/default.html
@@ -24,7 +24,7 @@
                 <li><a href="/upload">Upload</a></li>
             {% endif %}
             {% if current_user.can_view %}
-                <li><a href="/albums">Albums</a></li>
+                <li><a href="/album">Albums</a></li>
             {% endif %}
             {% if current_user.is_admin %}
                 <li><a href="/roles">Roles</a></li>
